trees: fix inheritence issues
If an instance of a subclass calls a super method, e.g.
FilesystemTree:delete calling self.super:delete, self in that case
becomes Tree and the fields and methods declared in FilesystemTree are not
available, i.e. TYPE and root in __tostring.

Define TYPE, root and current_node on Tree itself, by passing them in
the init method, instead of on the sublcasses. This is also "more
correct" as the fields are defined on Tree.