panels: enable 'search_for_node' for symbols and call hierarchy panels
panels: Add completion.on config for git status, symbols and call
hierarchy panels.
panels: Respect the completion.on config when searching search nodes in
the files panel.
nodes: Add an abbreviated path for LspDetailsNode which is based on the
root path and segmented with an arrow.
panels: Add method to TreePanel for completion using a list of
complete-items instead of just a list of strings.
panels: Use the abbreviated path when completing 'search_for_node' in
symbols and call hierarchy panels.